fix: split filter popover to satisfy max-lines ratchet
